Day 10
Python, #25/#9 today! I felt like I had a hard time wrapping my head around part 2, but I placed well so I guess everyone did. I took advantage of the fact that there were >= 200 unique directions (that is, that the laser doesn't make a full 360° loop) and didn't handle the other case (though I completed that case after the contest because why not). Not sure if the floating point is guaranteed to work out without small errors but the numbers were small so I crossed my fingers that it would.

NB: I did the coordinates in the opposite order from the problem description, so you need to remember to flip them (eg: `(8, 10)` means `1008` as an answer).