Day 13
Python, #25/#2!

Took a while to figure out what part 2 was saying. It would've been a lot clearer (to me) with a line like, "As the game proceeds, the output will reflect the new positions of the ball and paddle, and block tiles will be replaced with empty tiles. Your job is to find a series of joystick moves that cause all the blocks to be broken, at which point the program will terminate itself.", maybe with a link to Breakout too (not sure if I was supposed to draw on prior knowledge of the game (I did) or if the problem description was meant to be self-contained).

All my code does is move the paddle horizontally to match the ball's position. I was nervous that I might need to anticipate the ball's movement and have the paddle one step ahead, but thankfully that wasn't necessary.